fix(Material): rename file to index
