+++ src/views/Character/components/AscensionSummary/index.tsx
@@ -3,9 +3,9 @@ import { sortBy } from 'lodash';
 
 import {IAscensionSummary} from './types';
 import {IMaterial} from '../../../../charactersData/types';
-import Material from '../Material/Material';
 import {getCumulativeMaterials, getGroupedMaterials} from '../../../../utils/utils';
 import {COUNT, SORTING_INDEX} from '../../../../const/consts';
+import Material from '../Material';
 
 const AscensionSummary: FC<IAscensionSummary> = (props): ReactElement => {
   const cumulativeMaterials: IMaterial[] = getCumulativeMaterials(props)
